Możliwość dołączenia do gry

+++ NewWorld/Controllers/HomeController.cs
@@ -28,7 +28,7 @@ namespace NewWorld.Controllers
         }
 
         [Authorize]
-        public ActionResult GameList()  //jeżeli wartość id==true to nazwa jest zajęta
+        public ActionResult GameList()  
         {
             return View(GetHomeViewModel());
         }
@@ -67,9 +67,11 @@ namespace NewWorld.Controllers
         {
             ApplicationUser user = db.Users.Find(User.Identity.GetUserId());
             Game game = db.Games.Find(id);
+            UserGameProperty userGameProperty = db.UserGameProperties.Where(a => a.Player.Id == user.Id).Where(b => b.Game.Id == game.Id).FirstOrDefault();
+            if(game.IsBegan)
+                return RedirectToAction("GameList");
             game.Players.Remove(user);
-            UserGameProperty userGameProperty = db.UserGameProperties.Where(a => a.Player.Id == User.Identity.GetUserId()).Where(b => b.Game.Id == game.Id).FirstOrDefault();
-            game.UserGameProperties.Remove(userGameProperty);
+            db.UserGameProperties.Remove(userGameProperty);
             if (game.Players.Count == 0)
                 db.Games.Remove(game);
             db.SaveChanges();
@@ -83,7 +85,10 @@ namespace NewWorld.Controllers
             Game game = db.Games.Find(id);
             //jeżeli gra ma hasło wyswietl strone do wprowadzania hasła
             if (game.HavePassword())
-                return View(game);
+            {
+                JoinViewModel viewModel = new JoinViewModel { Id = id, Name = game.Name };
+                return View(viewModel);
+            }
             if (game.Players.Contains(user) || game.IsBegan)
                 return RedirectToAction("GameList");
             AddUserToGame(user, game);
@@ -94,14 +99,21 @@ namespace NewWorld.Controllers
         [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Join([Bind(Include = "Id,Password")] Game gameTmp)
-        { 
+        public ActionResult Join([Bind(Include = "Id,Password")] JoinViewModel viewModel)
+        {
+            Game game = db.Games.Find(viewModel.Id);
             if (ModelState.IsValid)
             {
                 ApplicationUser user = db.Users.Find(User.Identity.GetUserId());
-                Game game = db.Games.Find(gameTmp.Id);
+                if (game.Password != viewModel.Password)
+                    return View(new JoinViewModel { Id = viewModel.Id, Name = game.Name, WrongPassword = true });
+                if (game.Players.Contains(user) || game.IsBegan)
+                    return RedirectToAction("GameList");
+                AddUserToGame(user, game);
+                db.SaveChanges();
+                return RedirectToAction("GameList");
             }
-            return RedirectToAction()
+            return View(new JoinViewModel { Id = viewModel.Id, Name = game.Name, WrongPassword=true });
 
 
 
@@ -115,8 +127,9 @@ namespace NewWorld.Controllers
             var gameList = db.Games.Where(a => a.Players.Select(c => c.Id).Contains(user.Id)).ToList();
             //wybieranie tych w których jest aktywny
             gameList.Where(a => a.UserGameProperties.Where(b => b.Player.Id == User.Identity.GetUserId()).Where(c => c.Active).ToList().Count() == 1).ToList();
+            var gameListIds = gameList.Select(a => a.Id).ToList();
             //pobieranie otwartych gier z wyjątkiem tych do których dołączył gracz
-            List<Game> otherGames = db.Games.Where(a => !a.IsBegan).Except(gameList).ToList();
+            List<Game> otherGames = db.Games.Where(a => !a.IsBegan).Where(b=>!gameListIds.Contains(b.Id)).ToList();
             HomeViewModels viewModel = new HomeViewModels { YourGames = gameList, OpenGames=otherGames };
             return viewModel;
         }
